Rename ColorPicker to BaseColorPicker
