+++ frontend/src/lib/components/shared/ColorPicker/index.tsx
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { default } from "./ColorPicker"
+export { default } from "./BaseColorPicker"
